+++ lib_dil/ft_strmapi.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strmapi.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mvautrot <mvautrot@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ml <ml@student.42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 16:17:53 by mvautrot          #+#    #+#             */
-/*   Updated: 2022/12/21 14:30:22 by mvautrot         ###   ########.fr       */
+/*   Updated: 2022/12/30 10:50:31 by ml               ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/libft.h"
+#include "../includes/libft.h"
 
 char	*ft_strmapi(char  *s, char (*f)(unsigned int, char))
 {
-	size_t	i;
+	int	i;
 	char	*tab;
 
 	if (!s)
