+++ lib_dil/ft_substr.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mvautrot <mvautrot@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ml <ml@student.42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 12:28:59 by mvautrot          #+#    #+#             */
-/*   Updated: 2022/12/21 14:30:22 by mvautrot         ###   ########.fr       */
+/*   Updated: 2022/12/30 10:49:09 by ml               ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/libft.h"
+#include "../includes/libft.h"
 
-static char	*ft_slt(char  *s, size_t len)
+static char	*ft_slt(char  *s, int len)
 {
 	char			*dst;
-	unsigned int	i;
+	int	i;
 
 	i = 0;
 	dst = (char *) malloc ((len + 1));
@@ -30,7 +30,7 @@ static char	*ft_slt(char  *s, size_t len)
 	return (dst);
 }
 
-char	*ft_substr(char  *s, unsigned int start, size_t len)
+char	*ft_substr(char  *s, int start, int len)
 {
 	char			*dst;
